refactor: refactoring Contact component
